+++ pg_documentdb/src/operators/bson_expression.c
@@ -453,6 +453,8 @@ static MongoOperatorExpression OperatorExpressions[] = {
 	  FEATURE_AGG_OPERATOR_TOOBJECTID },
 	{ "$toString", &ParseDollarToString, &HandlePreParsedDollarToString,
 	  FEATURE_AGG_OPERATOR_TOSTRING },
+	{ "$toUUID", &ParseDollarToUUID, &HandlePreParsedDollarToUUID,
+	  FEATURE_AGG_OPERATOR_TOUUID },
 	{ "$toUpper", &ParseDollarToUpper, &HandlePreParsedDollarToUpper,
 	  FEATURE_AGG_OPERATOR_TOUPPER },
 	{ "$trim", &ParseDollarTrim, &HandlePreParsedDollarTrim,